make clusterIO use the refactored versions in cluster_directory
